Convert Sidekiq jobs to Active Jobs
