+++ app/jobs/hackathons/find_downed_websites_job.rb
@@ -1,0 +1,17 @@
+class Hackathons::FindDownedWebsitesJob < ApplicationJob
+  def perform
+    past_hackathons.find_each do |hackathon|
+      if hackathon.website_down?
+        hackathon.tag_with!("Website Down")
+      else
+        hackathon.untag("Website Down")
+      end
+    end
+  end
+
+  private
+
+  def past_hackathons
+    Hackathon.approved.where("ends_at < ?", Time.now)
+  end
+end
